# Return a wrapped span for SpanData so only the mutable data is copied.