This is a optimization of the creation of the SpanData implementation. 

master benchmark:
```
SpanPipelineBenchmark.runThePipeline_05Threads                                   thrpt    5  3352.511 ±  53.533  ops/ms
SpanPipelineBenchmark.runThePipeline_05Threads:·gc.alloc.rate                    thrpt    5  7233.932 ± 138.314  MB/sec
SpanPipelineBenchmark.runThePipeline_05Threads:·gc.alloc.rate.norm               thrpt    5  3400.001 ±   0.001    B/op
SpanPipelineBenchmark.runThePipeline_05Threads:·gc.churn.PS_Eden_Space           thrpt    5  7310.190 ± 695.826  MB/sec
SpanPipelineBenchmark.runThePipeline_05Threads:·gc.churn.PS_Eden_Space.norm      thrpt    5  3435.823 ± 315.746    B/op
SpanPipelineBenchmark.runThePipeline_05Threads:·gc.churn.PS_Survivor_Space       thrpt    5     0.232 ±   0.327  MB/sec
SpanPipelineBenchmark.runThePipeline_05Threads:·gc.churn.PS_Survivor_Space.norm  thrpt    5     0.109 ±   0.156    B/op
SpanPipelineBenchmark.runThePipeline_05Threads:·gc.count                         thrpt    5    79.000            counts
SpanPipelineBenchmark.runThePipeline_05Threads:·gc.time                          thrpt    5    50.000                ms
```

this branch benchmark:
```
SpanPipelineBenchmark.runThePipeline_05Threads                                   thrpt    5  3980.803 ± 253.972  ops/ms
SpanPipelineBenchmark.runThePipeline_05Threads:·gc.alloc.rate                    thrpt    5  6935.797 ± 455.120  MB/sec
SpanPipelineBenchmark.runThePipeline_05Threads:·gc.alloc.rate.norm               thrpt    5  2744.001 ±   0.001    B/op
SpanPipelineBenchmark.runThePipeline_05Threads:·gc.churn.PS_Eden_Space           thrpt    5  6984.225 ± 279.329  MB/sec
SpanPipelineBenchmark.runThePipeline_05Threads:·gc.churn.PS_Eden_Space.norm      thrpt    5  2763.550 ± 139.847    B/op
SpanPipelineBenchmark.runThePipeline_05Threads:·gc.churn.PS_Survivor_Space       thrpt    5     0.245 ±   0.330  MB/sec
SpanPipelineBenchmark.runThePipeline_05Threads:·gc.churn.PS_Survivor_Space.norm  thrpt    5     0.097 ±   0.135    B/op
SpanPipelineBenchmark.runThePipeline_05Threads:·gc.count                         thrpt    5    71.000            counts
SpanPipelineBenchmark.runThePipeline_05Threads:·gc.time                          thrpt    5    46.000                ms
```

Or graphically (from a different run): 

Here are the benchmark results from master:  

![image](https://user-images.githubusercontent.com/858731/82482474-c1c0ef80-9a8b-11ea-897a-5fea04a5a085.png)

And on this branch:

![image](https://user-images.githubusercontent.com/858731/82482495-cdacb180-9a8b-11ea-9b38-0c17af8d62c7.png)