@jkwatson 2020-05-15
> Can you expand on the idea where are you not happy?

Mostly, I'm not sure that this solution will have much of an impact in the number of allocations. Do you have a benchmark you've been using so we can compare?