@jkwatson 2020-05-19
> I don't but we can write one to confirm. I did look only at a running up and got a snapshot of heap allocs

I'll see if I can cook something up today.