@jkwatson 2020-05-19
ok. I added a benchmark to this PR. I ran this against both master and this branch. This branch was about 10-12% faster, and had a similar percentage of decrease in allocations.