@bogdandrutu 2020-05-19
And I am confident that we can optimize even more :)