@bogdandrutu 2020-05-20
@jkwatson please add in the description results for the benchmark.