@jkwatson 2020-05-20
> Please rebase to pick the codecov change :)

I think this needs some unit tests on the wrapper, which I hadn't written yet.