@jkwatson 2020-05-20
rebased (also, I had locally added the unmodifiable wrappers and it didn't significantly change the benchmarks, so they slipped in here, as well).