@jkwatson 2020-05-20
> @jkwatson my bad I did not expect that you added the change, it is not correct right now. I will create a PR to fix this.

What's the issue? I can fix it, since I'm in the code adding test coverage.